requirements.txt
discord
docutils
discord.py
asyncio
python-dotenv
mcstatus